requirements.txt
aiogram==2.12.1
aiohttp==3.7.4.post0
aioschedule==0.5.2
APScheduler==3.7.0
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
gino==1.0.1
idna==2.10
lxml==4.6.2
multidict==5.1.0
numpy==1.20.2
pycryptodome==3.9.9
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.24
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
yarl==1.6.3
